refactor: moved LRU cache to shared
