+++ packages/next/src/server/lib/router-utils/filesystem.ts
@@ -14,7 +14,7 @@ import path from 'path'
 import fs from 'fs/promises'
 import * as Log from '../../../build/output/log'
 import setupDebug from 'next/dist/compiled/debug'
-import { LRUCache } from '../lru-cache'
+import { LRUCache } from '../../../shared/lib/lru-cache'
 import loadCustomRoutes, { type Rewrite } from '../../../lib/load-custom-routes'
 import { modifyRouteRegex } from '../../../lib/redirect-status'
 import { FileType, fileExists } from '../../../lib/file-exists'
