+++ packages/next/src/server/require.ts
@@ -9,7 +9,7 @@ import { normalizePagePath } from '../shared/lib/page-path/normalize-page-path'
 import { denormalizePagePath } from '../shared/lib/page-path/denormalize-page-path'
 import type { PagesManifest } from '../build/webpack/plugins/pages-manifest-plugin'
 import { PageNotFoundError, MissingStaticPage } from '../shared/lib/utils'
-import { LRUCache } from '../server/lib/lru-cache'
+import { LRUCache } from '../shared/lib/lru-cache'
 import { loadManifest } from './load-manifest.external'
 import { promises } from 'fs'
